Bubble

+++ hdr/menu.hpp
@@ -33,7 +33,7 @@ private:
     sf::Texture mBackgroundTexture;
     float mBackgroundCounter = 0;
     float mBackgroundCounter2 = -1280;
-    uint8_t mColorCounter = 0;
+    double mColorCounter = 0.0;
     Button mPlay; 
     Button mControls;
     Button mSkins;
